requirements.txt
cryptography==3.4.8
pyOpenSSL==21.0.0
